WIP GetTeamStats

+++ src/Smab.TTInfo.TTLeagues/Models/Group.cs
@@ -3,9 +3,9 @@
 internal sealed class Group
 {
 	public List<Match> Matches { get; set; } = new();
-	public long Type { get; set; }
+	public int Type { get; set; }
 	public DateTimeOffset? Date { get; set; }
-	public long Range { get; set; }
-	public long? Week { get; set; }
+	public int Range { get; set; }
+	public int? Week { get; set; }
 	public required string Title { get; set; }
 }
