+++ src/Smab.TTInfo.TTLeagues/Models/MatchPlayer.cs
@@ -4,5 +4,5 @@ internal sealed class MatchPlayer
 {
 	public required string UserId { get; set; }
 	public required string Name { get; set; }
-	public long? Score { get; set; }
+	public int? Score { get; set; }
 }
