+++ src/Smab.TTInfo.TTLeagues/Models/MatchTeamInfo.cs
@@ -2,18 +2,18 @@
 
 internal sealed class MatchTeamInfo
 {
-	public long Id { get; set; }
+	public int Id { get; set; }
 	public required string UserId { get; set; }
-	public long? TeamId { get; set; }
+	public int? TeamId { get; set; }
 	public required string Name { get; set; }
 	public required string ShortName { get; set; }
 	public List<Member>? Members { get; set; }
-	public long? Score { get; set; }
+	public int? Score { get; set; }
 	public List<object>? Reserves { get; set; }
-	public long Type { get; set; }
-	public long? Points { get; set; }
+	public int Type { get; set; }
+	public int? Points { get; set; }
 	public required string Captain { get; set; }
-	public long? ClubId { get; set; }
+	public int? ClubId { get; set; }
 	public required string CaptainId { get; set; }
 	public required string DisplayName { get; set; }
 }
